Deprecate AADB2CJwtBearerTokenAuthenticationConverter, use AADJwtBearerTokenAuthenticationConverter instead.
