+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/aad/webapi/AADJwtBearerTokenAuthenticationConverter.java
@@ -2,28 +2,35 @@
 // Licensed under the MIT License.
 package com.azure.spring.aad.webapi;
 
+import com.azure.spring.aad.AADJwtGrantedAuthoritiesConverter;
 import com.azure.spring.aad.AADOAuth2AuthenticatedPrincipal;
-import com.azure.spring.aad.AbstractJwtBearerTokenAuthenticationConverter;
 import com.azure.spring.aad.implementation.constants.AADTokenClaim;
 import com.azure.spring.aad.implementation.constants.AuthorityPrefix;
 import org.springframework.core.convert.converter.Converter;
+import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.OAuth2AccessToken;
 import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.BearerTokenAuthentication;
+import org.springframework.util.Assert;
 
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
 
 /**
  * A {@link Converter} that takes a {@link Jwt} and converts it into a {@link BearerTokenAuthentication}.
  */
-public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerTokenAuthenticationConverter {
+public class AADJwtBearerTokenAuthenticationConverter implements Converter<Jwt, AbstractAuthenticationToken> {
+
+    private final Converter<Jwt, Collection<GrantedAuthority>> converter;
+    private final String principalClaimName;
 
     /**
-     * Construct AADJwtBearerTokenAuthenticationConverter by AADTokenClaim.SUB and DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.
+     * Construct AADJwtBearerTokenAuthenticationConverter by AADTokenClaim.SUB and
+     * DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP.
      */
     public AADJwtBearerTokenAuthenticationConverter() {
         this(AADTokenClaim.SUB, AADResourceServerProperties.DEFAULT_CLAIM_TO_AUTHORITY_PREFIX_MAP);
@@ -31,6 +38,7 @@ public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerT
 
     /**
      * Construct AADJwtBearerTokenAuthenticationConverter with the authority claim.
+     *
      * @param authoritiesClaimName authority claim name
      */
     public AADJwtBearerTokenAuthenticationConverter(String authoritiesClaimName) {
@@ -39,6 +47,7 @@ public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerT
 
     /**
      * Construct AADJwtBearerTokenAuthenticationConverter with the authority claim name and prefix.
+     *
      * @param authoritiesClaimName authority claim name
      * @param authorityPrefix the prefix name of the authority
      */
@@ -55,17 +64,28 @@ public class AADJwtBearerTokenAuthenticationConverter extends AbstractJwtBearerT
      */
     public AADJwtBearerTokenAuthenticationConverter(String principalClaimName,
                                                     Map<String, String> claimToAuthorityPrefixMap) {
-        super(principalClaimName, claimToAuthorityPrefixMap);
+        Assert.notNull(claimToAuthorityPrefixMap, "claimToAuthorityPrefixMap cannot be null");
+        this.principalClaimName = principalClaimName;
+        this.converter = new AADJwtGrantedAuthoritiesConverter(claimToAuthorityPrefixMap);
     }
 
     @Override
-    protected OAuth2AuthenticatedPrincipal getAuthenticatedPrincipal(Map<String, Object> headers,
-                                                                     Map<String, Object> claims,
-                                                                     Collection<GrantedAuthority> authorities,
-                                                                     String tokenValue) {
-        String name = Optional.ofNullable(principalClaimName)
-                              .map(n -> (String) claims.get(n))
-                              .orElseGet(() -> (String) claims.get("sub"));
-        return new AADOAuth2AuthenticatedPrincipal(headers, claims, authorities, tokenValue, name);
+    public AbstractAuthenticationToken convert(Jwt jwt) {
+        OAuth2AccessToken accessToken = new OAuth2AccessToken(
+            OAuth2AccessToken.TokenType.BEARER, jwt.getTokenValue(), jwt.getIssuedAt(), jwt.getExpiresAt());
+        Map<String, Object> claims = jwt.getClaims();
+        Collection<GrantedAuthority> authorities = converter.convert(jwt);
+        OAuth2AuthenticatedPrincipal principal = new AADOAuth2AuthenticatedPrincipal(
+            jwt.getHeaders(), claims, authorities, jwt.getTokenValue(), (String) claims.get(principalClaimName));
+        return new BearerTokenAuthentication(principal, accessToken, authorities);
+    }
+
+    private static Map<String, String> buildClaimToAuthorityPrefixMap(String authoritiesClaimName,
+                                                               String authorityPrefix) {
+        Assert.notNull(authoritiesClaimName, "authoritiesClaimName cannot be null");
+        Assert.notNull(authorityPrefix, "authorityPrefix cannot be null");
+        Map<String, String> claimToAuthorityPrefixMap = new HashMap<>();
+        claimToAuthorityPrefixMap.put(authoritiesClaimName, authorityPrefix);
+        return claimToAuthorityPrefixMap;
     }
 }
